requirements.txt
streamlit
torch
pandas
scikit-learn
pillow
pytesseract
